debt - btoa/atob is not core
