+++ src/vs/base/test/common/strings.test.ts
@@ -417,9 +417,4 @@ suite('Strings', () => {
 	test('getGraphemeBreakType', () => {
 		assert.equal(strings.getGraphemeBreakType(0xBC1), strings.GraphemeBreakType.SpacingMark);
 	});
-
-	test('multibyteAwareBtoa', () => {
-		assert.equal(strings.multibyteAwareBtoa('hello world'), strings.multibyteAwareBtoa('hello world'));
-		assert.ok(strings.multibyteAwareBtoa('平仮名'));
-	});
 });
