+++ src/vs/workbench/browser/parts/editor/textDiffEditor.ts
@@ -31,7 +31,7 @@ import { CancellationToken } from 'vs/base/common/cancellation';
 import { EditorActivation, IEditorOptions } from 'vs/platform/editor/common/editor';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 import { isEqual } from 'vs/base/common/resources';
-import { multibyteAwareBtoa } from 'vs/base/common/strings';
+import { multibyteAwareBtoa } from 'vs/base/browser/dom';
 
 /**
  * The text editor that leverages the diff text editor for the editing experience.
